@@ -7,7 +7,7 @@ dependencies {
     implementation  group: 'com.google.auto.service', name: 'auto-service', version: '1.0-rc6'
 }
 
-sourceCompatibility = "7"
-targetCompatibility = "7"
+sourceCompatibility = "8"
+targetCompatibility = "8"
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
