@@ -1,11 +1,13 @@
 package com.bty.retrofit.demo.log.FloatingWindow;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.ListView;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import com.bty.retrofit.demo.R;
 
 public class DebugLogActivity extends AppCompatActivity implements LogManager.LogAddListener {
